+++ client/src/models/Recipient.js
@@ -3,7 +3,7 @@
 // models/
 // Recipient.js
 
-import { Peer } from './index.js';
+import { FileStore, Peer } from './index.js';
 
 /* Recipient - Receive a file
  *
@@ -24,6 +24,9 @@ import { Peer } from './index.js';
  * 'close' -> Event
  * - The connection was closed. Will reconnect if file isn't received.
  *
+ * 'bytes' -> MessageEvent (event.data has bytes got.)
+ * - An update on the number of bytes got.
+ *
  * 'download' -> MessageEvent (event.data: { href, name, size, type }.)
  * - The file was downloaded
  */
@@ -31,77 +34,186 @@ export class Recipient extends EventTarget {
   // private fields
 
   _peer = null; // the peer connection
-  _builder = null; // the file builder (Web Worker)
-  _dc = null; // the data channel
-  _blob = null; // the file as it's being built
+  _dataChannel = null; // the data channel
+
   _fileinfo = {}; // the information about the file being received
+
+  _fileStore = null; // the file store
+  _request = null; // a file store request
+  _blob = null; // the current blob
+  _offset = 0; // the current offset
 
   // constructor
 
   constructor(id) {
     super();
 
+    // open a peer connection with the dropper
     this._peer = new Peer(id);
 
     // on data channel negotiated
     this._peer.addEventListener('datachannel', (event) => {
-      this._dc = event.channel;
-      this._dc.addEventListener('message', this._onDcMessage.bind(this));
+      this._dataChannel = event.channel;
+      this._dataChannel.binaryType = 'arraybuffer';
+      this._dataChannel.addEventListener(
+        'message',
+        this._onDataChannelMessage.bind(this)
+      );
     });
 
     // on WebRTC peer connected disconnected
     this._peer.addEventListener('disconnected', () => {
-      this._dc = null; // old data channel is dead
+      // old data channel is dead; wait for renegotiation
+      this._dataChannel = null;
+    });
+
+    // open the file store database
+    this._fileStore = new FileStore();
+
+    this._fileStore.addEventListener('error', (event) => {
+      console.log(`Recipient: Got error in file store: ${event.data}`);
+    });
+
+    // on file store opened
+    this._fileStore.addEventListener('open', () => {
+      this._request = { readyState: 'pending' };
     });
   }
 
-  _onDcMessage(event) {
+  async _clearFile() {
+    await this._fileStore.clearFile(this._fileinfo.name, this._fileinfo.size);
+
+    // ready for blobs
+    this._request = { readyState: 'done' };
+  }
+
+  _addBlobToFileStore() {
+    // add the current blob to the file store
+    this._request = this._fileStore.addBlob(
+      this._fileinfo.name,
+      this._offset,
+      this._blob
+    );
+
+    this._request.addEventListener('error', (event) => {
+      console.log(
+        `Recipient: Got error in _addBlobToFileStore: ${event.target.error.toString()}`
+      );
+    });
+
+    // update offset and blob
+    this._offset += this._blob.size;
+    this._blob = null;
+
+    this.dispatchEvent(new MessageEvent('bytes', { data: this._offset }));
+  }
+
+  async _download() {
+    // get the file
+    let file = await this._fileStore.getFile(
+      this._fileinfo.name,
+      this._fileinfo.size,
+      this._fileinfo.type
+    );
+    this._fileStore.close(); // close the file store
+
+    // create download link for file
+    const href = URL.createObjectURL(file);
+
+    file = null; // garbage
+
+    // dispatch event with download information
+    this.dispatchEvent(
+      new MessageEvent('download', { data: { href, ...this._fileinfo } })
+    );
+  }
+
+  async _onDataChannelMessage(event) {
     if (typeof event.data === 'string') {
       const message = JSON.parse(event.data);
 
       switch (message.type) {
         case 'fileinfo':
           this._fileinfo = message.fileinfo;
+
           this.dispatchEvent(
             new MessageEvent('fileinfo', { data: message.fileinfo })
           );
 
+          // file store is waiting for us to consider the file info
+          if (this._request !== null) {
+            this._clearFile();
+          } else {
+            this._fileStore.addEventListener(
+              'open',
+              this._clearFile.bind(this)
+            );
+          }
+
           break;
 
         case 'done':
-          this._dc.send('{"type":"received"}');
+          this._dataChannel.send('{"type":"received"}');
 
           // close everything; drop is complete
 
-          this._dc.close();
-          this._dc = null;
+          this._dataChannel.close();
+          this._dataChannel = null;
 
           this._peer.close();
           this._peer = null;
 
-          const href = URL.createObjectURL(
-            new Blob([this._blob], { type: this._fileinfo.type })
-          );
+          if (this._request.readyState === 'pending') {
+            try {
+              // wait for this pending request to finish
+              await new Promise((resolve, reject) => {
+                this._request.addEventListener('success', resolve);
+                this._request.addEventListener('error', reject);
+              });
+            } catch (err) {
+              console.log(
+                `Recipient: Got error when wrapping up: ${err.toString()}`
+              );
+            }
+          }
 
-          // free this reference
-          this._blob = null;
+          if (this._blob !== null) {
+            // add this blob to the file store
+            this._addBlobToFileStore();
 
-          // dispatch event with download information
-          this.dispatchEvent(
-            new MessageEvent('download', { data: { href, ...this._fileinfo } })
-          );
+            try {
+              // wait until finished
+              await new Promise((resolve, reject) => {
+                this._request.addEventListener('success', resolve);
+                this._request.addEventListener('error', reject);
+              });
+            } catch (err) {
+              console.log(
+                `Recipient: Got error when wrapping up: ${err.toString()}`
+              );
+            }
+          }
+
+          // start download; will dispatch 'download' event when ready
+          this._download();
 
           break;
       }
     } else if (event.data instanceof ArrayBuffer) {
       if (this._blob !== null) {
-        // build blob with this new part
+        // append the message to the existing blob
         this._blob = new Blob([this._blob, event.data]);
       } else {
+        // create a new blob with the message
         this._blob = new Blob([event.data]);
       }
+
+      // check if the last request is done
+      if (this._request !== null && this._request.readyState === 'done') {
+        this._addBlobToFileStore();
+      }
     } else {
-      console.log(`Got unexpected data: ${event.data}`);
+      console.log(`Recipient: Got unexpected data: ${event.data}`);
     }
   }
 }