+++ client/src/models/index.js
@@ -5,6 +5,7 @@
 
 export * from './DataChannel.js';
 export * from './Dropper.js';
+export * from './FileStore.js';
 export * from './Peer.js';
 export * from './Recipient.js';
 export * from './SignalChannel.js';